Fix for MB-17783 - don't show query workbench when connected to nodes without a query service.
Due to limitations in the UI plug-in architecture, the query workbench can only function if the
browser is connected to a node that is running the query service. This change checks for the presence
of a query service when the UI is loaded in the browser. If there is no query service, an error
message is shown. If there exist other nodes in the cluster running the query service, a link
to those nodes is shown. If there is a query service, the workbench runs normally.

Change-Id: I84509d1326ee9069a83a1451c53dddcb79032a93
Reviewed-on: http://review.couchbase.org/59416
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>