Fix for MB-18360, limit the amount of data that the workbench tries to show.
The browser has trouble rendering large amounts of data. The Ace
editor gets pretty slow at around 10MB, and the table and tree
visualizers get really slow with more than about 750KB of data. The
latter is not too surprising when you consider that all the data
elements are rendered as HTML TDs and LIs. Since Javascript is
single-threaded, the process of rendering large amounts of data brings
the browser to a standstill, preventing the user from doing anything.

This fix sets limits (in bytes) on the amount of data that it will try
to render in Ace (around 10MB) or the tree/table views (around
750KB). If the data set is larger, instead of trying to render it, the
browser shows a message suggested that the user select a lower limit
value.

The user can still save the full dataset to a file. Having it in
memory doesn't affect the browser too much, but trying to render it
does!

Change-Id: I9da2515edfe8943654a2da33f53ccebdf09051c4
Reviewed-on: http://review.couchbase.org/60589
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>