Fix for MB-18193, autocompletion issues in the query UI.
The main issue of the bug was autocompletion of elements in subobjects
inside arrays, but while working on that I discovered errors when
autocompleting subobjects, not to mention overzealous matching of
terms in autocompletion. For the longer term, too many of the tweaks
to autocompletion were in the generic Ace code
(ext-language-tools.js), instead of in the n1ql specific code
(mode-n1ql.js), making it harder to track updates to the Ace code
editor. I've managed to extract almost everything into mode-n1ql (with
the exception of two small items that couldn't be extracted). Just to
test, I updated Ace to the latest version, and removed all the
unnecessary modes to reduce clutter.

Change-Id: I19b921b171c480edc95e77259ad41dca31ed9f52
Reviewed-on: http://review.couchbase.org/60892
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>