Remove the "Limit" text field in the UI, add size limiting in result displays.
Originally, the "Limit" text field in the Query Workbench was added so
that every query would be automatically limited in size, to prevent
results that would freeze the browser during rendering. Yet this
automatic limit feature caused some problems of its own, since
wrapping the user's query in additional N1QL sometimes could break the
user's query. It also made errors more difficult to understand, since
the limit wrapping could change the location of a syntax error.

As a better solution, the automatic limit has been removed, but the
results displays will not automatically show large results, and
instead show a warning message. The message suggests using a lower
limit, or a more restrictive 'where' in the N1QL query. If the user
doesn't care about having their browser frozen for the time required
to render the result, they can click a link to cause the rendering to
proceed.

Change-Id: I0ac38cf7b923b6eddc4d9242db4311bbc95c6f7a
Reviewed-on: http://review.couchbase.org/61109
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>