MB-18621 - fix problem with queries with semicolons.
The problem is that angular was taking my "Content-Type:
application/json" and urlencoding the parameters and changing the
content-type to application/x-www-form-urlencoded. Unfortunately,
angular uses a weird version of urlencoding that leaves semicolons in
place. This messes up the server parsing of parameters - golang's
parsing assumes that a semicolon marks the end of the parameter.

The fix was to pre-urlencode the query, with an extra replace to make
sure the semicolons get encoded. Surely there's a better way, but I
couldn't find out how to force angular to keep my http request as json
encoded.

Change-Id: Iee3d8a54994746069649635709b0c41b7db8e080
Reviewed-on: http://review.couchbase.org/61721
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>