Use new getUrlsRunningService function.
In addition to removing the hardcoded path for the link generation
this change fixes a bug where a node with the query service is added
to a cluster, but rebalance hasn't run yet. This has the effect
of providing a link to the node that when clicked brings you to a
page on that node that again provides a link to itself. We could
allow users access to the query service on inactiveAdded nodes (as
the query service is functional by that point) but that's a separate
conversation.

Change-Id: Ic228ebe62d882e9e08f96d290e5d9065b2a6b314
Reviewed-on: http://review.couchbase.org/71614
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>