MB-22640 - The problem was a document containing HTML in either the field names, values, or document IDs would have that HTML executed by the Tree or Table views. These views did not use Angular data binding (which prevents this sort of thing) for efficiency reasons - Angular was just not able to handle reasonable-sized data tables.
For a solution, I investigated Angular's $sanitize service, but it
only works properly on valid HTML, it will throw an exception on HTML
fragments (which might be found in pathalogical data). The current
solution is to replace '<' and '>' with '&lt;' and '&rt;' That
eliminates all HTML tags, so there is nothing to induce behavior in
the browser.

Change-Id: Ic6f287edcbf611e54dab513816463bc5ed9d4a2a
Reviewed-on: http://review.couchbase.org/73473
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>