MB-23122 - Improve the error messages when there is no contact with the query engine. If there is no engine on this node, say as much. If the engine is supposed to be there but is not available, give an appropriate message.
When the user doesn't have permission to do monitoring, don't show the
monitoring tab. Also give an appropriate error if they navigate
manually to that page.

Possible fix for MB-23648, which may have been due to improperly
cached data. Automatically check the query engine status, available
buckets and permissions whenever the user logs in, so a change of
users gets appropriate permissions.

Also change the visible statistics layout if the user does not have
permission to retrieve the chart statistics.

Change-Id: If3c1c2be0a65c64ee77eee7ba465943c772d3f40
Reviewed-on: http://review.couchbase.org/76447
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>