Bring back the fixed width for inner cells, since nested tables were not displaying properly.
Add an option in the code for either scrollbars or shadows, so we can
switch back and forth depending on design priorities.

Bring back scroll bars and a nested div for inner scrolling, since
otherwise in some cases the screen ends up completely blank.

Work toward adding a bottom shadow, though it's not yet good enough.

Change-Id: Ibb0273d46b650d19257f75243f04f0edd71b38cc
Reviewed-on: http://review.couchbase.org/85547
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>