Fix left-alignment issue with datatable, so that column headers and data are aligned on the left.
Setting the left-padding for data-table-cell to the same value was
causing problems in some cases, due the extra padding in the
data-table-cells that wrapped arrays and objects. The fix involved
changing the directive to use a non-padding style for cells containing
arrays and objects. Also, there was a second alignment issue for some
column headers because their caret span had not been moved to the
right of the label.

Change-Id: Ia61f874ed60c9871d103639bf246a4859454a39e
Reviewed-on: http://review.couchbase.org/86532
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>