Improvements to handling of large documents in new doc editor.
The tabular view now truncates at 500 columns, to keep the table from
getting too large. In that case a warning is shown at the top. Users
can still edit the full document in the text editor. I tested the text
editor up to 20MB documents, and performance was acceptable.

Also fix a race condition where sometimes the index status of a bucket
was not known when it came time to retrieve the first batch of
documents. Now the index status is retrieved at the same time as the
bucket names.

Change-Id: Ibe0868b9414db36692395fe2266691c77e6d0f8e
Reviewed-on: http://review.couchbase.org/88348
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>