MB-29118 and MB-29122 - issues with very long numbers.
When copying existing documents, it was missing a check for long
numbers, and thus copying a parsed version of the document, which
loses precision for long numbers.

For creating new documents (MB-29122), the issue was that the regex
for detecting long numbers did not detect negative numbers. When
fixing this I discovered that it did not correctly detect floating
point numbers either, and both those issues are now fixed.

Change-Id: I1d5deb026765d014e8f3e860876b1b4ff8e2355b
Reviewed-on: http://review.couchbase.org/92269
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>