MB-29572 - query result processing very slow on large data sets.
On exceptionally large datasets containing numbers too long to
represent in javascript (such as 'select * from `travel-sample`')
using the js_beautify() call on the client to format the JSON can take
a long time - over a minute. Changed back to having formatting occur
on the server, which is much faster.

Also, the fix_long_number service was taking unneccesarily
long. Removed the count of long numbers, since we only need to know if
there is one.

Change-Id: I7547708f7106355b57961e7f44d8ac2299adacd4
Reviewed-on: http://review.couchbase.org/93961
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>