Many small fixes to the new Doc Editor.
MB-29619 - unnecessary reload when hitting return on entering text to
search the bucket menu. Problem was that the menu was unnecessarily
part of an HTML form with the other controls.

MB-29627 - not correctly displaying documents that contain bare
strings or arrays in "simple" mode. Addition of simple mode didn't
account for this condition.

MB-29630 - allow edit dialog to adjust layout when error message shown.

MB-29628 - number fields in spreadsheet mode needed to allow any
number of decimal places.

MB-29622 - button order was incorrect for binary documents.

MB-29621 - change representation of empty array from "[]" to "[ ]" to
make it easier to read on the screen.

MB-29620 - boolean fields with value 'false' were not showing a widget
in the spreadsheet view.

MB-29618 - avoid duplicate error message if copy fails due to already
existing document.

MB-29613 - improve the error messages when the user forces an invalid
value into the offset or limit fields.

MB-29609 - fix the misalignment between the size of the spreadsheet
table and the size of the header.

MB-29607 - add the style 'overflow: hidden' to prevent very long field
names from overlapping the next column in the spreadsheet view.

Change-Id: I019cbf805574b2a876e814b349bf9b8a3c21f473
Reviewed-on: http://review.couchbase.org/94180
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>