MB-29624 - problem with copy results as tab-separated.
The process of creating tab-separated data requires an array of
objects which get flattened into a table. If we don't have an array of
objects, as with an empty result, or the result of an EXPLAIN query,
it was not able to create tab-separated values, and thus the clipboard
was unchanged.

Now an error message is shown if tab-separated data cannot be created.

Change-Id: Ie61f9d2d1796234e23f9d3e4f7341c45090a01c3
Reviewed-on: http://review.couchbase.org/94313
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>