MB-29400 - Handling very large documents in the new doc editor.
With certain very large documents, in certain browsers (Chrome &
Firefox), the regex check for very large numbers (> 2^53) was causing
an out-of-memory exception. Add code for properly handling exceptions
in this case. N.b. an exception means that we don't know if the
document includes very large numbers, so tabular editing is
disallowed.

Change-Id: I508ca64d583e673ad1942e5e1edf048ec6b2a844
Reviewed-on: http://review.couchbase.org/94484
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>