MB-29748 - Show count queries per minute, not average.
In the query monitoring page, in the stats panel at the bottom, the
indicators for "queries > xxx ms" were showing the *average* of the
per-second values for the past minute. That isn't terribly
useful. Instead we are now showing the *sum* for the past minute, so
the user can see exactly how many long-running queries have been going
on recently.

Change-Id: Ib190279d2092de9262af7ca0e7871683f8f2b17d
Reviewed-on: http://review.couchbase.org/94607
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>