MB-29356 - Display Analytics query plans as DAGs when needed.
Analytics query plans can effectively be DAGs. The JSON description is
actually a tree, but various plan nodes are repeated when they are on
multiple paths.

Unfortunately D3 does not support DAGs, only trees. I was able to
simulate a DAG, however, but forcing second and subsequent appearances
of multiply-occuring nodes to appear in the same location as the
original. This actually works in the cases I tried. I did have to add
some CSS logic to hide duplicate edges, however, because they were not
showing up correctly.

Change-Id: I9b0630f40d47d76ad5964b83a8c535842dbd9487
Reviewed-on: http://review.couchbase.org/96235
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>