MB-30232 - hide documents tab for users with only query select.
A user with only query select permissions is not able to retrieve
documents via the REST API, the server fails and an HTTP 500 error is
returned. This is because the user doesn't have permissions to read
the XATTRs, which are returned with the document via the REST API. See
also MB-30232.

This fix hides the "Documents" tab if the user does not have
permissions to read XATTRs.

Change-Id: Ice78069fd184fde3854d463bf36b62653e0c0ee6
Reviewed-on: http://review.couchbase.org/96459
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>