MB-30669 - fix race condition for very fast queries and timings
The Query Workbench automatically runs 'explain' for every query, so
that a query plan can be shown while the query is running. If the user
has selected "Collect Query Timings", when the query completes, it is
replaced with the plan including query timings.

When running a very fast query, such as:

select * from `travel-sample` use keys ["airline_10748"];

Sometimes the 'explain' takes longer than the query itself, meaning
that the 'explain' plan overwrites the query timings plan.

This change fixes that race condition.

Change-Id: I8683d1d1f23ed54ac692ee0ec4e146255313e3c5
Reviewed-on: http://review.couchbase.org/97464
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>