MB-30639 - better 'enter' handling in QWB preferences
The problem reported was that clicking 'enter' in *any* of the text
fields in the preferences dialog would cause a numbered parameter to
be added. Apparently with angular, 'enter' in any text field in a form
causes the first button in the form to be clicked, even if a later
button is marked as 'type="submit"'. The solution is to somewhat
redundantly label each of the buttons as 'type="button"' (except for
the 'ok' button, which is type="submit").

Change-Id: I9f502370de4bdb14ee5c2f0676f8e10607567a9f
Reviewed-on: http://review.couchbase.org/97891
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>