MB-30598 - Handling timeouts and other errors in the QWB
The query workbench was sending a timeout parameter to ns_server,
which proxies requests to cbq-engine, but cbq-engine also has a
timeout parameter. Change the code to send the same timeout value to
both. This means that queries interrupted by the timeout are marked as
"timeout" instead of "completed", because cbq-engine times out instead
of the proxy gateway.

When cbq-engine times out, however, it is possible to have partial,
truncated results from the server. Thus I also had to improve the
notifications in the query workbench so that warnings and/or errors
are shown along with partial results, if any.

Change-Id: Icae57d6e1f387d7709c3dfcfe5d7ed871cb74a6c
Reviewed-on: http://review.couchbase.org/98027
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>