MB-30080 - performance issues in new doc editor
The new doc editor was slow when dealing with very complex documents
(many fields). This is caused by the browser spending time rendering
many many text-entry areas, one for each field. There's no general way
to speed that up (though Safari is about twice as fast as Chrome in
these situations), but I was able to help this two ways:

1) use ng-if instead of ng-show to prevent rendering of the
spreadsheet text-entry areas when the user selects "simple" mode. This
makes the simple mode effectively as fast as the classic doc editor.

2) there was already a limit on the maximum number of fields to show,
which was set at 500. This existed to prevent the browser from
grinding to a halt when facing thousands of fields. I reduced this
number to 250, which speeds up the "spreadsheet" mode somewhat. And if
the user really needs to edit more than 250 fields, they can use the
"edit as JSON" functionality.

Change-Id: Ief83b5018c3f39fbe7a269fb49dc6379ff7c9c8b
Reviewed-on: http://review.couchbase.org/98136
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>