MB-31448 - handle doc-not-found errors better
In the new doc editor, documents are retrieved using the REST
API. When retrieving an invalid document key, or a document that is
deleted during the query, the server returns a 404 error, "Object Not
Found". That error message is not ideal, so we detect the 404 and use
"Document not found" instead. We also don't show the error dialog,
since the error will also show up in the results section.

Also fix a problem with the ID/Range setting not being saved and
restored.

Change-Id: I97429034fb2f419d2a41c6345ef69aa9eb584014
Reviewed-on: http://review.couchbase.org/101348
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>