CBSE-5825 - Run multiple queries
MB-17825 - Allow pasting of multiple queries at once

This change is the first step in resolving the above issues. Now, when
multiple queries are pasted in, the queries are executed in order,
with the execution ceasing if any query fails. Execution works
properly at this stage, as does cancelling a batch of queries.

The current limitations are:

- the query editor needs to support SHIFT-enter to permit addition of
  carriage returns after semicolons to prevent execution after the
  first query.

- The results of all the batched queries are concatenated, which is
  not idea. At some point we will need a better way of tabing or
  paging between separate results.

- The query plan text is also concatenated, and the query plan viz
  doesn't work at all.

Change-Id: I23d6a596df1ddd9e7bb29764e0f30b7693d208b5
Reviewed-on: http://review.couchbase.org/102940
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>