CBSE-4626 - edit documents larger than 256kb
The server now accepts requests up to 20MB in size. There must be some
overhead, because saving a 13MB document causes an error with the
request exceeding 20MB. I have thus set the maximum size for document
editing to 10MB, to be safe. Even at that size it is pretty slow.

Change-Id: Ib65e574412cc410f6c709da588a159722feb58a4
Reviewed-on: http://review.couchbase.org/106644
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>