Improve handling of cases with slow INFER.
The Query Workbench runs INFER on each bucket. If you have a bucket
with huge documents, or many large documents, INFER can take a long
time, and result in a schema that is impossible for the browser to
display. This change, combined with some improvements to INFER on the
server side, change the settings to speed up INFER, and handle the
very large schemas that sometimes result.

Change-Id: I33f0a28196281f98c2dd0a6988609e52329795c5
Reviewed-on: http://review.couchbase.org/106904
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>