MB-33307 - Maximize usable vertical space in QWB
The UI now has click handlers to detect whether the user is clicking
on the query editor, or on the results part of the page. If the user
is clicking on the query, that gets prioritized for space, whereas if
the user is clicking on the results, that gets more space.

Change-Id: I21cad4d6173274e99e576e5cedd4de0787b3726b
Reviewed-on: http://review.couchbase.org/106972
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>