MB-33707 - Add option to not save queries to local storage.
Saving the workbench query history to local storage could be
problematic on shared machines. Of course people should never use
shared machines anyway because of the possibility of keystroke
loggers, but if anybody still complains, this change includes an
option which prevents the UI from saving the query history to the
browser's local storage.

Change-Id: I43a497a04093f4cdae31886eb5d7a5accafb2504
Reviewed-on: http://review.couchbase.org/108030
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>