MB-32946 - Display issues in QWB schema panel
With the default behavior, a slice of text in the panel was being
shown to the right of the scrollbar. This could be fixed for Sarafi
and Chrome, but Firefox insists on drawing the scrollbar that way, and
doesn't offer any way to change it. I tried adjusting spacing many
different ways, but Firefox wouldn't budge.

Change-Id: Ic0aef952901c0b848c8f900b9ddedb15c611ebce
Reviewed-on: http://review.couchbase.org/111943
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>