MB-35191 - Partial fix, doc editor handling docs with id "." or ".."
The doc editor would hang when trying to show a list including
documents with the id "." or ".." (because those have an overloaded
meaning in URLs). While urlencoding the ID works for calls to the REST
API using curl, it doesn't work from the GUI (Angular's $http treats
%2E the same as ".").

This fix improves the error handling so that the doc editor doesn't
hang, and any documents with the ID "." or ".." are shown with an
error message. Also, trying to add a document from the GUI with one of
those IDs fails. Still, the source of the problem is now visible, and
users can try N1QL if they really want to manipulate such documents.

Change-Id: I665b5e7e0f731435794df7b8bef24842124dea0c
Reviewed-on: http://review.couchbase.org/112435
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>