MB-35798 - reduce the number of calls to the query service for metadata
Prior to this change, the qwQueryService was making many REST calls to
the query service to get the metadata for displaying in the query
workbench, and for use in the document editor. For initially testing
authorization on a bucket, it would call "select count(*) from
<bucket>", and if that succeeded it would proceed to get the inferred
schema. After that, every 5 seconds it would issue a "select count(*)"
query for each bucket to keep the document counts up to date.

With this change, the schema inference call is used to test
authorization on a bucket, and a single N1QL query is used to retrieve
the document count for every bucket. This reduces the number of
initial REST calls to 2 + #buckets. $timeout is also used to slow down
the rate of initial metadata queries.

Change-Id: Ideab6828dbd68c381c4deb0ed34343d5cca33ebc
Reviewed-on: http://review.couchbase.org/114566
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>