MB-35850 - update table and tree views when multiple queries.
The directives for table and tree views use the angular $watch to keep
track of when the underlying data changes. Pushing new results onto
the results array, however, wasn't recognized as a change, however,
because the top-level array didn't change. With this change, we clone
the result array, so that the directives notice the change. (The
alternative was to tell the directive to do deep equality, but that
could be very expensive with large results.)

Change-Id: Ida3b068df167adfeefd2700bf58a98f62d7eebc4
Reviewed-on: http://review.couchbase.org/114634
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>