MB-35857 - Auto-shrink too aggressive
Right now the query workbench has a variable indicating whether the
user is focused on the query, or on the results, and expands the
relative sizes of those parts of the UI accordingly. This change makes
the value of that variable persistent when you go to other screens. So
if you click on Documents, then return, your focus will stay the same.

This change also changes the focus if the user scrolls either the
query or JSON results. I.e., if the user scrolls the query, and it is
not focused, it will become focused and enlarge, and vice versa for
the JSON results.

Change-Id: I87607e965fea1bbe73efa37a00f407b0df3ad4a3
Reviewed-on: http://review.couchbase.org/114729
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>