MB-35987 - incorrect message after error contacting query service
Pre-MadHatter, the Query Workbench was only available when connected
to a node running the query service. When an error occurred, it
offered a list of nodes on which the query service was running. In
some circumstances this list was inappropriate: e.g., when the query
nodes are warming up, and thus aren't available regardless of node. We
didn't have a good way to detect the difference, however, so the list
was shown regardless.

With the advent of proxying in MadHatter, the workbench should be
available from any Couchbase node. As a result, it is no longer
appropriate to show the list of nodes running the query service. This
change removes that list, and instead highlights the error that
occurred when trying to contact the query service, and suggests
reloading the browser in case the query nodes are warming up, or
suffering some other transient error.

Change-Id: Ie185ae72c3c18af02e85f5cce4089bed56e48f9b
Reviewed-on: http://review.couchbase.org/114832
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>