MB-36537 - no way to cancel running query
Whenever a query is run, the query workbench might in background also
run "EXPLAIN" and "ADVISE" on the query to provide the user with more
information. These usually take only a fraction of a second, much less
than most queries. In javascript, the query, EXPLAIN, and ADVISE are
all run as separate $http requests, and $q.all is used to wait for the
requests to finish. $q.all will finish immediately if any of the
promises is rejected (returning with an error of any kind). EXPLAIN
and ADVISE very seldom fail, usually the problem is with the regular
query, so this hasn't been a problem up to now.

In the case of this bug, a large query result from a remote cluster
with slow network bandwidth was taking several minutes to return, and
ADVISE was returning an error state because DP mode was turned
off. This led the $q.all to finish, yet the query was still running
and hadn't returned its result. The solution is to make sure each
$http.then error function returns Promise.resolve() to allow $q.all to
continue until everything is done.

Change-Id: Ic3628c784e93f3fb491a4e69b822c8d286d7d5c0
Reviewed-on: http://review.couchbase.org/116796
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>