Show correct query status when 'Explain' results in error.
In the query workbench, 'Execute' invokes code paths for both a query
Explain and query Execute. When both paths are being executed (in
parallel), the query execution path should provide the overall query
status at the end, since the explain path is supposed to be a
behind-the-scenes thing. When the user clicks 'Explain' however, only
the 'Explain' path is followed, and that query status should be used.

Before this change, if the explain-only path was being followed, the
resulting status was being used in the success case, but not the error
case. Now the explain status is being used for both cases.

Change-Id: I485579cdabe326e616b26398d975e984dda4f68e
Reviewed-on: http://review.couchbase.org/120430
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>