MB-37942 - Query Workbench slow with very large bucket schemas.
The bug reported a performance regression with Query Workbench going
from 6.0 to 6.5. The problem arose from very large schemas on the
supportal cluster (> 1MB). The schema display directive was creating
the HTML for the entire schema display even if the flavor is not shown
on the screen. The fix uses 'ng-if' to prevent creation of the schema
HTML for flavors that are not expanded.

Change-Id: I1a9dc333fa340f38f12ba0abddc273fb88e40ec5
Reviewed-on: http://review.couchbase.org/122553
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>