MB-38333 - Web-based data import, found in the Documents tab.
This is the initial check-in of the code for the web-based data import
feature. While about 10x slower than cbimport, it is aimed at new
users who are playing around with Couchbase, and those with small data
sets.

This provides a nice usability advantage over cbimport, since it
automatically detects the file type, allows selection of key fields
from a menu, and allows the user to avoid the complex syntax of
cbimport. It does not support keys comprised of multiple fields,
however. In those cases there is always cbimport.

Import works by: 1) loading a data file into memory in the browser, 2)
guessing the data format using a regular expression, 3) parsing the
data, 4) uploading the data by creating a series of N1QL batch INSERT
commands (keeping them less than 5MB to avoid any issues with proxying
on ns_server).

Import is limited to files of 100MB or smaller, because the browser
gets really slow with files that size. A future improvement could be
to read the data file in chunks, but that significantly increases the
complexity of parsing the data, and at that point why not use
cbimport.

Change-Id: I665c4816d174fa4469883fc0e4b9cb50351bb930
Reviewed-on: http://review.couchbase.org/c/query-ui/+/126356
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>