MB-38596 - For the CE version, add link to online advisor service
Couchbase's website offers an online query advisor service. Since the
query advisor is not supported in CE, add a link from the CE query
workbench to open a tab showing the online query advisor. Ideally the
online advisor would support parameters, so the workbench could pass
the query as a parameter, but that is not yet supported.

Change-Id: I1de6e3afbc43b223182886db8a0827c3a1e65d84
Reviewed-on: http://review.couchbase.org/c/query-ui/+/127896
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>