Add a version directory entry for UI compat version 7.0
The UI framework has changed in an non-backward compatible way in
Cheshire Cat and we should change the UI compat version to reflect
that. This patch adds a compat version directory for version 7.0.
Once all pluggable components have a version directory for 7.0
we can change the current UI compat version and remove the entry
for the 5.0 directory.

Change-Id: I5e985072e3a78a2c84822f94a94be37ec21cc5af
Reviewed-on: http://review.couchbase.org/c/query-ui/+/127934
Tested-by: Dave Finlay <dave.finlay@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>