MB-39400 - Sorting was broken in the document editor table view.
Sorting of the data columns in the tabular view of the doc editor was
broken, and probably had been broken for a long time (ever since the
the textual view was added as an alternative to the tabular view). The
original approach was to add sort listeners into the header elements,
but since the header of the tabular elements could be hidden, the
listeners didn't work. The solution was to be smarter and use ng-click
along with some $event magic to get the pointer to the header element
that was clicked on.

The same bug highlighted an issue with "show top keys" which was
sort-of a mode, but not really. These changes make it a real, separate
mode from regular document queries, so the user can toggle back and
forth.

Change-Id: I97a31dcb50bff38bdac567a990e30650a0448ad8
Reviewed-on: http://review.couchbase.org/c/query-ui/+/129229
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>