MB-39903 - numbers imported as strings in CSV, TSV MB-39911 - crash when specify improper field for doc ID MB-39910 - incomplete cbimport command MB-39899 - can't import arbitrary JSON file MB-39905 - docs with duplicate Doc IDs will overwrite other docs MB-39912 - show error when there is no query service
MB-39903 - the D3 libary used for parsing CSV & TSV treats numbers as
strings unless a flag is passed asking it to import numbers when
possible. This change includes that flag.

MB-39911 - the UI was showing incorrect fields for use as DocID, and
if the user selected such a field the import would fail. The fix
prevents showing fields that are not present in every document.

MB-39910 - two options were missing from the generated cbimport
command, one due to a typo, the other due to an omission.

MB-39899 - cbimport only supports two subsets of JSON: a list of
documents, and a text file with one JSON document per line. This bug
reports that arbitrary JSON files can't be imported. This is not a
surprise. The fix is to change the wording and add a tooltip to help
inform the user that arbitrary JSON is not supported.

MB-39905 - this bug reports that importing documents with the same Doc
ID as existing documents causes the existing documents to be
overwritten. This is expected behavior to those familiar with
Couchbase, so the fix is to add a tooltip explaining the meaning of
document IDs.

MB-39912 - the import UI relies on the query service to work. This fix
shows an error message when there is no query service in the
cluster. The user can still specify an import, and get the generated
'cbimport' command, but they can't click the 'import' button.

Change-Id: I33c1196a9beb8c84abe591fc966cb0acbea3001e
Reviewed-on: http://review.couchbase.org/c/query-ui/+/130666
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>