Migrate Document Editor to use Angular 8
This change brings the Document Editor into the exciting new world of
Angular 8. In order to support the migration, I had to port a variety
of directives and services which should speed porting other parts of
the UI.

There is also a "Hello World" component which demonstrates the
simplest possible component to fit in our UI. It is also useful for
testing migrated directives and services.

Change-Id: Ibae7d5d1e740f34891ab9d5ab820d3ca391f53c6
Reviewed-on: http://review.couchbase.org/c/query-ui/+/132201
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>