MB-42400 - poor workbench performance when many collections
The query workbench was using the query-engine for retrieving buckets,
scopes, and collections. When the number of collections is very high
(> 500), the query-engine is very slow in returning this metadata,
sometimes taking up to 5 minutes. This change shifts over to using the
REST API to retrieve bucket, scope, and collection information.

Change-Id: Ica30f817c5f0cc6e3d87948db3a5a472fa8df071
Reviewed-on: http://review.couchbase.org/c/query-ui/+/140141
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>