MB-42308 - problem with options specified in the options dialog
Part of MB-42308 was that none of the options specified in the options
dialog were being included in the http request, nor were they saved
between sessions. The problem was due to the qw_query_controller,
which is AngularJS, trying to access the option structure in
qw.query.service, which is Angular 8. Apparently the downgrade wrapper
prevents this from working properly, any direct changes in the
AngularJS world are only visible there. The solution is to use an
accessor function, which does allow the changes to propigate between
the two Angular worlds.

Change-Id: I29e20c238e4a587b63ae33cbf44b2b85cfeac7c5
Reviewed-on: http://review.couchbase.org/c/query-ui/+/140145
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>