MB-42653 - get positional and named parameters working again
There was one more aspect to this bug - the change in Angular meant
that HTTP interceptor was changing the 'Content-Type' of HTTP requests
from 'application/json' to urlencoded. This was preventing the
positional parameters from working properly. I fixed the syntax,
making sure to use the new Angular HttpHeaders, and now the parameters
work properly.

Change-Id: Ib9af326b66b0626d26d43953be70ef87a8789428
Reviewed-on: http://review.couchbase.org/c/query-ui/+/140154
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>