MB-42715 - Problem displaying digital documents in doc editor
An earlier MB requested that binary documents be displayed not as
Base64, which is non-human-readable, but as decoded Base64. This helps
in the case where human-readable text is present in the binary data,
but in the case of ATR documents the binary data contains a null
character, which javascript treats as the end of string. The solution
to this problem is to run the binary data through JSON.stringify after
decoding, which produces a human-readable form. E.g., the null
character becomes "\u0000".

Change-Id: Ibc258bfbed7e36c13b821cfce2237da248a487b3
Reviewed-on: http://review.couchbase.org/c/query-ui/+/141707
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>