Fix a problem detecting query types with multi-line queries.
The query workbench automatically runs EXPLAIN and ADVISE in the
background for normal queries, so it needs to detect whether a query
is a normal SELECT versus EXPLAIN, ADVISE, PREPARE, etc. The regular
expressions were set to use multiline mode, which meant that they
might detect a "SELECT" statement starting a line other than the
first.

This also fixes an issue where errors in batch queries were sometimes
improperly handled.

Change-Id: I2c7002576bbf16a22c33e9dd33511dfeaac06e5a
Reviewed-on: http://review.couchbase.org/c/query-ui/+/141716
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>