MB-43244 - Don't run advise on explain queries and vice versa
The query workbench automatically runs explain and advise on regular
queries so the user can understand why the query performed the way it
did and how to improve it. It's not supposed to run advise on explain
queries, however, nor vice versa.

Also add hooks for running queries as implicit transactions, so that
the UI can add some affordance for doing so. (MB-43089)

Change-Id: I8727ac6235746495b9aaa1c8cd51a455c34c5f6b
Reviewed-on: http://review.couchbase.org/c/query-ui/+/141992
Tested-by: Eben Haber <eben@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>